qt: use QWheelEvent::angleDelta() instead of QWheelEvent::orientation()
It's deprecated in Qt 5.15 [^1], since 5.0 [^2], and removed in Qt 6.
Although the deprecation was delayed to 5.15 because of a Qt3D issue [^3], that
we don't use.
We use a replacement is suggested in [^1].

The commit log of [^2] even mentions its deprecated "since Qt4 times".

The code is slightly similar to what is found in qtWheelEventToVLCKey().
And borrows from 7498ac55db73dec778d97de88a758707964ebdc7.

[^1]: https://doc.qt.io/qt-5/qwheelevent-obsolete.html#orientation
[^2]: https://github.com/qt/qtbase/commit/7d29807296cb7ccc7f3459e106d74f93a321c493
[^3]: https://github.com/qt/qtbase/commit/8dc45d5b7cd9637f28b2d872a834266b991bf434