qt: use QMouseEvent::globalPosition() instead of QMouseEvent::globalX/Y()
globalPosition() doesn't exist in Qt5 but globalPos() is the same thing [^2].

It's deprecated in Qt 5.15 [^1], since 5.0 [^3], and removed in Qt 6.
Although the deprecation was delayed to 5.15 because of a Qt3D issue [^4], that
we don't use.
We use a replacement is suggested in [^1].

The commit log of [^3] even mentions its deprecated "since Qt4 times".

[^1]: https://doc.qt.io/qt-6.2/qmouseevent-obsolete.html#globalX
[^2]: https://doc.qt.io/qt-5/qmouseevent.html#globalPos
[^3]: https://github.com/qt/qtbase/commit/7d29807296cb7ccc7f3459e106d74f93a321c493
[^4]: https://github.com/qt/qtbase/commit/8dc45d5b7cd9637f28b2d872a834266b991bf434