qt: add possibility to disable anti-aliasing in SDFAARoundedTexture.frag
Although this adds the possibility, due to lacking build system support
I do not manually create a new file that defines `ANTIALIASING` but
rather use it unconditionally in both `SDFAARoundedTexture.frag` and
`SDFAARoundedTexture_cropsupport.frag` and override the antialiasing
property as readonly true.

When the build system supports defines for the shaders, we can start
respecting `antialiasing` and revert the readonly attribute.