qt: introduce `TextureProviderItem`
An invisible item that acts as a texture provider. It
uses `QSGTextureView` under the hood, and allows defining
a viewport on the source item (which needs to be texture
provider itself) with the property `textureSubRect`.

This is similar to `sourceRect` with layering. However,
this consumes no extra resources unlike layering where
it needs to allocate a frame buffer object and do offscreen
rendering (even though the source is a texture itself, it
is still not necessary).