qml: set alpha to zero in Noise.frag to use additive blending
Since the output color is pre-multiplied, setting the alpha to
zero does not mean that nothing is shown. Rather, the alpha is
already factored in the RGB channels so it is respected.

As the Qt docs says, we can set the alpha to 0.0 to achieve
additive blending (S + D) that is necessary for noise effect.