win32: Prevent VLC from registering as the default player for MIDI files
This commit resolves an issue where the installation of VLC on Windows
unintentionally designated it as the default application for MIDI files.
Given that Windows supports MIDI playback natively through Windows Media
Player and that VLC requires a SoundFont for proper MIDI playback, this
change ensures that VLC does not disrupt the default MIDI file associations,
maintaining the expected user experience.

Fixes: #27464