qml: do not let qt to create implicit layer in artist top banner
Here, Qt wants to create a layer because when source is `QQuickImage`,
`QGfxSourceProxy(ME)` checks its fill mode and source size.

In this case, we need to tell Qt that we don't need a layer here, as
we are doing the mapping as we want but Qt currently does not offer
a way to tell that and simply tries to decide itself: "If needed,
MultiEffect will internally generate a ShaderEffectSource as the
texture source"

Using the static texture for the blur source has also the advantage
of applying blur only once because the source would not get updated
constantly due to size change. With clipping, we can simply adjust
the viewport without needing to update the blur.

Unfortunately we can not do this with older Qt versions, because if
the source image changes fill mode from stretch or to stretch where
it is fed to a blur effect, the effect either causes crash (layering
needed to not needed, so Tile -> Stretch) or does not work (layering
not needed to needed, so Stretch -> Tile).