qt: restore window backdrop blur effect in CompositorX11
CompositorX11, unlike others, requires explicit code to
adjust backdrop blur. When the effect gets activated at
a later time after initialization, it currently does not
have code to do the required adjustment.

This is a regression since b1e11641. I did not notice
it because I was using extended frame with CSD, where
this regression is masked.