qt: do not use access for resource or local images
Access was meant to be used for remote images as we stopped
using Qt Network, Qt can not retrieve remote images.

QQuickImage can handle local or resource files better for
its needs. It does not need to use an image provider for
local files.

VLCAccessImage is not good when same images loaded together.
When I add an album to the playlist, even though the items
share the same image (url), they appear to be loaded multiple
times. I assume this has something to do with VLCAccessImage
being an asynchronous provider.

With the default behavior, as long as cache is enabled, the
images are re-used when possible.