qt: use `QWindow::setVisible()` instead of `QWindow::show()`
`QWindow::show()` does not necessarily care about the state
of the window, but rather uses the default behavior for the
platform.

Since we adjust the window geometry and states before the
window is shown (which is the expected behavior, as otherwise
the window would get adjusted immediately after shown), using
`QWindow::show()` potentially overrides the window states
nullifying the behavior of the interface starting maximized.