qml: disable add displaced transition and clear add transition artifact in `ListViewExt`
Qt has number of bugs that hinder using transitions properly. The worst one seems to be
add displaced, which causes items to have incorrect positions, where it is not possible
to recover from that situation.

Transition in general causes visual glitches (independent of positioning), but it seems
to be possible to recover from that by adjusting the viewport in an undetectable way.

This can be observed when there are 2-3 items in the playlist, and when a large number,
such as 30 items are inserted at the beginning of the playlist.

Instead of disabling the transitions altogether, I propose to:
1) Clear the artifact on items after items are added.
2) Disable add displaced transition, as it is not possible to recover from incorrectly
positioned delegate once it occurs. Keep move and remove displaced transition