qt: do not use viewporter protocol if Qt does not use fractional scaling
Viewporter protocol is used for fractional scaling, and currently we do
not use the fractional scaling protocol ourselves because Qt uses it and
provides scale through the interface window.

However, Qt starts supporting fractional scale on Wayland starting with
6.5.0. If Qt version is lower than 6.5.0, it is not necessary to use the
viewporter protocol because Qt is not going to provide the fractional
scale anyway.

I don't think it is a good idea to start using the fractional protocol
ourselves to get the fractional scale, instead of Qt providing that. The
reason is the video window should use the same scale as the interface
window for video integration, and it complicates the code unnecessarily.

I have not realized this before, because I test Qt code with Qt 6.2 and
Qt 6.8, and I can not test Wayland changes on Qt 6.2 because it has never
been working for me on Wayland.