dbus: multiple fixes related to mpris:trackid URIs
Fix org.mpris.MediaPlayer2.TrackList.TrackAdded including the same
mpris:trackid value to both Metadata and AfterTrack fields. Now it
correctly sets the right path when the track is added to the start of
the playlist.

Use playlist item id instead of its index during format and parsing of
mpris:trackid URIs. This will avoid confusion in the event tracks are
replaced or moved around.