qt: reflect rate change immediately
If the change ends up being different, then
it can be corrected later when the timer
provides the update.

The user should not wait couple of seconds
for the timer to update the rate for normal
cases. For position, it is interpolated but
for the rate I propose using the requested
rate as is.