opengl: fix chroma check from UpdateFormat()
The fmt passed to vlc_gl_interop_New() is const, the new changed chroma
is in interop->fmt_in.

This fixes crashes and assert after vd->ops->update_format() that
required a chroma change.

PS: if update_format() fails, a video converter will be inserted in the
static filter chain.