fourcc: specify color_bits
This will be used with subtype to calculate a possible color loss when
converting between 2 chromas.

It's needed for GPU chromas and the few chromas where the pixel_size and
plane description are not enough to calculate the bits per color.

Example:
    { VLC_CODEC_ABGR,                  PACKED_FMT(RGB, 4, 32, 8) },
    { VLC_CODEC_RGBA10LE,              PACKED_FMT(RGB, 4, 32, 10) },

    { VLC_CODEC_VUYA,                  PACKED_FMT(YUV444, 4, 32, 8) },
    { VLC_CODEC_Y210,                  PACKED_FMT(YUV422, 4, 32, 10) },

Same buffer size but different color size.

It's an average for the few chromas with different color size, like
RGB565. In that case, color_bits = (5 + 6 + 5) / 3. It is also possible
to add a new color description array in the vlc_chroma_description_t for
a more precis color size definition.