vout_subpicture: fix "fit inside" SPU computation
The restrained size is in source unscaled coordinates
when apply_scale is not set.
We need to unscale the output dimensions to check if it fits inside in that case.

Fixes positioning of SPU's when the window is smaller than the source.

Plus rename SpuAreaFitInside() to SpuAreaMoveInside() as it's not fitting, ie
stretching anything, just moving the subpicture (and probably shouldn't for
absolute placement anyway).