opengl: renderer: add support for video-stereo-mode
The video-stereo-mode is setup the same way I did the projection mode.
The client needs to restart the renderer to re-compile the shaders and
re-setup the variables for the mode, avoiding the need for a specific
API. This is justified by changing video-stereo-mode not happening at
each frame.

Note that projection_mode is temporarily enforced to
PROJECTION_MODE_RECTANGULAR whenever the multiview_mode is enforced to
another value than the source, to avoid trying to change the projection
from a texture containing the two eyes, leading to weird behaviour. This
happens when setting the SIDE_BY_SIDE stereoscopic mode which is made
for showing the two frames, but also in the STEREO stereoscopic mode
which is not yet supported in the module (need vlc-vr work).