emscripten: extra_checks: ban incompatible function casts
Emscripten is known to deal badly with function casts to incompatible
types [^1]. This patch prevent us to accidentally merge harmful UB casts
to wasm enabled code. The warning is really strict but is compliant to the
standard. Code that wants to use function cast should be disabled from
the wasm platform.

[^1]: https://emscripten.org/docs/porting/guidelines/function_pointer_issues.html