contribs: png: disable debug lib suffix for all platforms
Given contribs are bootstrapped with `--disable-optimizations` option,
the CMAKE flag `-DCMAKE_BUILD_TYPE` is set to `Debug`, and when
build configuration is set to `Debug` in libpng contrib, it CMAKE
configuration append a `d` char to the static library name,
resulting in a lib named `libpng16d.a` instead of `libpng16.a`.
Then some vlc module depending on libpng aren't able to link
properly as they can't find their lpng16 dependency.