qt: fix queued task leak and use after free in ThreadRunner
As the serial task queue doesn't autodelete. When the task was canceled while
running, this caused the serial task to be leaked.

Moreover, the TaksRunner wasn't traking these tasks explicitly, the serial task
may try to acces the TaskRunner after it was deleted.

Now we assing a task ID to serial tasks to track them as the regular ones.