qml: get rid of `showArtist()` signal in `MusicArtistsAlbums.qml`
It is not clear why such signal exists, it seems that it is
supposed to be a function instead.

Regardless, having such signal is not meaningful because there
is already `currentIndex` which is synchronized with the shown
artist. This synchronization is satisfied by the change handler.

The change handler already needs to adjust the artist id (for
synchronization), making `showArtist()` meaningless.

Currently clicking a delegate does not change the current index.
Views should change the current index when clicking on delegate.