d3d9_fmt: link with d3d9.lib
The Direct3D9 code is enabled for non-UWP targets and all Windows targets
have this DLL (since Vista).
If for some reason the DLL is not there, the plugin won't be loaded
and it will rightfully be skipped.

Direct3DCreate9Ex() is available since Vista [^1]:

> Direct3DCreate9Ex is supported only in Windows Vista, Windows Server 2008, and Windows 7. Earlier versions of the D3D9.dll library do not include Direct3D9Ex and Direct3DCreate9Ex.

We fallback to Direct3DCreate9() just in case it fails for maximum compatibility.

[^1]: https://learn.microsoft.com/en-us/windows/win32/api/d3d9/nf-d3d9-direct3dcreate9ex#remarks