dvdnav: fix DemuxForceStill()
Wait for the es_out to be empty/drained in order to avoid flushing past
video/subtitles.

This also fixes an undefined behavior (that could lead to no audio) when
DemuxForceStill() was called but the PCR was not reset (cf. the other
call to DemuxForceStill() not showing in this diff).