macosx: use correct variable names for actual arch
This is used in other places in the script and if unset leads to the
`-arch` argument being passed without any architecture, which is invalid
and causes compiler invocations to fail.

The reason this worked in CI is that the build script overwrites the
ACTUAL_HOST_ARCH as it can be changed by an argument. However the
env.build.sh script is used also outside the build.sh script in
manual build scenarios, which is the whole reason it is a separate
script to begin with.