decoder: check if draining from IsEmpty()
Temporary commit that restore the behavior from 0dd8096f46a9. This
allows demuxers to check if empty without draining the decoders. This
behavior i undefined as this could lead to audio underrun.

This is needed to no break bisect while fixing the dvdnav module. This
will be reverted just after.