d3d11: link directly with d3dcompiler.lib
It redirects to the d3dcompiler_47.dll in mingw-w64 and in the Windows
SDK that's the only variant there is. This is also how D3DCompile()
is supposed to be used [^1].

The DLL is available in Windows 7 with Platform Update [^2].
No version of d3dcompiler_XXX.dll is available in vanilla Windows 7
which doesn't have Direct3D11 anyway.

[^1]: https://learn.microsoft.com/en-us/windows/win32/api/d3dcompiler/nf-d3dcompiler-d3dcompile
[^2]: https://learn.microsoft.com/en-us/windows/win32/direct3darticles/platform-update-for-windows-7