contrib: qt: disable sbom generation
This was recently made enabled by default. Until
all contribs are ready for SBOM, it should not be
necessary for Qt to generate SBOM.

Note that it appears different Qt packages might
not respect the option applied for Qt base, so
we still need to add "sbom" to the ignored list
at the moment.