qt: probe rhi asynchronously and cache the valid result for the next run
Startup time is defined as the time it takes to start playing
the initial item when the application is opened.

Currently probing is done synchronously each time the interface
starts. This is not ideal, as we don't expect the system to
suddenly start supporting a particular graphics api. Obviously,
due to hardware change or driver update or any reason this may
change, so we still need to check that each time.

In this case, the worst can happen is that `QQuickWindow`
emits error and terminates the application (usually with an
error message box) when the api is no longer supported.
However, since the cached api is checked each time asynchronously,
the next startup would be fine.

This basically improves the startup performance, at the expense
of causing an error and terminating the application (in the
worst case) if the system suddenly starts not supporting once
valid api. The worst case is a rare case, so I believe that
this would be a good optimization.