vout_intf: setup video-stereo-mode callback in init
Dummy video outputs only define variables and don't use the interactive
interface callbacks as they are not running. Because the callback was
set during the initial variable setup, it would lead to assertion inside
the video_output.c code as the stereo mode was changed dynamically from
a dummy video_output object.