macos: fix media playback
Given playback was stopped, many issues happened when the playback
was resumed from the main controls bar.

The media time/progress slider wasn't visible anymore and app
crashed unexpectdly when picture in picture was activated.

This was caused by the `_currentMedia` var that was set to `nil`
when playback is stopped but never reassigned when playback was
resumed.

This change will reassign the var with the current played back input
item from the vlc player when the `currentMedia` getter is called.