dvdnav: use the dvd playtime as PCR
Using the PS SCR value as a PCR isn't correct in the DVD use case. DVDs
are made of multiple non-continuous PS streams that aren't constrained
to have a consistent SCR across streams.
Basically, each VOB changes might trigger a PCR reset. This was fine in
VLC 3, but now the PCR is used to drive the clock and the seek bar which
leads to hazardous behaviors, the most noticeable now is the seekbar
incorrect timestamps.

This patch uses the DVD playtime as the PCR and adapts the DTS instead.
This ensure a continuous PCR curve even on VOB changes.