caf: don't keep extra data for an unknown codec
The codec ID is set in ReadDescChunk(). And according to the error before using it:

> The audio description chunk must be the first chunk

We should not be using codec data if we don't have the codec.
The file will eventually fail with this error, but we don't need to
copy bogus data before that.

Fixes #29012

(cherry picked from commit b3ec88d53167d4a3383db43328f10abcc454cebe)