qml: take exclusive grab in CSD title bar tap handler
We need to take exclusive grab here, otherwise handlers
beneath that take exclusive grab may act on press.

This is the case when media library view items go under
the title bar. Unfortunately in controls we have to use
`TapHandler.ReleaseWithinBounds` as gesture policy for
the tap handlers due to a bug present in Qt 6.2.