modules: meson: always store module manifest
The module manifest is indexed by the name and allows checking whether a
given module has been declared or not. In particular, it's used by the
tests to check whether the module was enabled or not. For this feature
to work, it should always be stored and not just when the module is
actually enabled.