contrib: qtbase: add patch to treat queued connections fairly on windows
On Windows, Qt does not treat queued connections fairly. This is a problem
for us, because most of the operations we are doing are asynchronous, for
optimization or utility reasons.

We also use delayed bindings in complex layouting, such as within the player
control bar where controls need to be invisible when they no longer fit into
the space (unless expandable).

In summary, this seems to how Qt behave with regard to queued connections:

- In the worst case (messages are constantly pushed), they may be delayed
forever if Windows keep sending messages. This is the case with window
resize, for example.

- In the average case (there were messages when event dispatcher
encountered with a queued call), they are delayed by 10~ milliseconds.

- In the best case, there is no delay, but this is rare for a GUI
application, as GUI applications almost always have messages due to
user interaction.