vout_subpicture: handle the full window placement in video per region
"subtitle in black bars" is not opt-in per SPU codec, and per region
for each codec. It is off by default.

Each subpicture needs to decide for itself if it's using the whole display area.

We can't tell before the region is created/updated if it's going to be in video coordinates
or display coordinates.

The core doesn't need to know anymore if a picture is a "subtitle"
(contains timestamps) in "absolute" (video or screen?) coordinates.
It just needs to provide both dimensions.