configure: test protobuf-lite < 22 with minimum/maximum values
When the protoc version is in the a.b.c format the protobuf version
is in the a.b.c.d format. Some version of pkg-config allow the equality
of the two, but not all.

We can just make sure protobuf is between a.b.c and a.b.(c+1).

Fixes #29066