meson: rust: always define dependencies from rust
To be able to declare modules unconditionally, declare the rust
dependencies and state variables that will be used when declaring
modules, without executing the commands that depend on Rust.

In short, this allows using the enabled: flag instead of conditionally
defining the module under an if get_option('rust').allowed() block.