input: decoder: trace input fifo buffer level
The decoder fifo level metric allows interpreting the situation that led
to the fifo from the decoder to be flushed (ie. the fifo reached 400MiB
of undecoded data stored). The metrics generates a lot of points though,
so be sure to aggregate when displaying it.

Example of query in flux language:

    from(bucket: v.defaultBucket)
    |> range(start: v.timeRangeStart, stop:v.timeRangeStop)
    |> filter(fn: (r) =>
      r._measurement == "measurement" and
      r._field == "fifo_size"
    )
    |> aggregateWindow(every: 1s, fn: max)

The size metrics are casted into uint64_t, but are always below the
uint64_t max value anyway given they represent lesser values than
400MiB.