demux: subtitle: check the FORMAT= value is a number in mpsub
We don't need vlc_strtof_c() since we already have the float value.
We also don't need to check what is after the number (a comment for example).
As long as it's a number it should be FPS. Anything else other than
"FORMAT=TIME" is bogus. So we don't need to get the remaining part of the string.

(cherry picked from commit a2c276959052b592fb37eeed0968ffeac02f0014) (edited)
edited:
- us_strtof has not been renamed to vlc_strtof_c
- sub-original-fps is called sub-fps on 3.0