qml: gracefully handle the case when image is not available in `ImageExt`
- Use root size for the shader if image is not available, as in
  that case the image would not have a sensible size. The implicit
  size is chosen arbitrarily to be 64x64 instead of 0x0, which
  is mainly relevant when root does not have an overridden size.
- Do not bother calculating the crop rate if there is no image.
- Do not do outlining if there is no image, as in that case it
  means there is nothing to outline.

This ensures that background is shown while the image is loading.

`ShaderEffect` uses a transparent dummy texture with repeat wrap
mode when the texture is not available. With source over blending
of the texture into the background, a fully transparent source
yields the destination color, which is the background color.