qml: rename `TableViewDelegate.qml` to `TableViewDelegateExt.qml`
Qt 6.9 introduces `TableViewDelegate` in Qt Quick Controls, this
causes name conflict here as the import is not explicitly stated
when resolving `TableViewDelegate` in `TableViewExt.qml` currently.

Alternatively the name can be kept, and `Widgets.TableViewDelegate`
can be used to specify which `TableViewDelegate` to use, but this
is probably safer than that approach.