qt: handle `QVariant` in `MainCtx::urlListToMimeData()`
Sometimes the array elements may be variant (`QVariant`),
even though the actual type is a `QUrl` or a `QString`.

This is an odd case because `QUrl` is automatically
converted to `url` value type when passed to QML. However,
the source url may already be provided as a `QVariant`,
where in that case this conversion would not occur.