qml: make header color transparent by default in `TableViewExt`
Having a header color by default does not make sense here:
- There is header background color but not background color for the rest.
- Views are not controls, and are often placed within controls where
background is handled properly.
- A purely transparent rectangle does not create a scene graph node thus
not rendered. This is considered an optimization for rendering. In this
case, since the background is likely to be the same as the background
color of the parent control, the area is painted twice with the same
color. If the color needs to be different, the property can be overridden
to deviate from transparent color.