qt: do not initialize unused qt quick effects shaders
As per Qt Quick Effects' `CMakeLists.txt`,
the excluded resources are the ones that
use masking.

We do not use mask, so we don't need to
initialize these shaders. This should save
some disk space for the static build.

We could eliminate more of these, for example
depending on the blur level, only a certain
subset of these shaders are used. However,
I decided to not be that aggressive for the
moment.